fix bottom panel resizing
